optimized tour activity

+++ app/src/main/java/com/balinasoft/clever/model/TickHandler.java
@@ -1,5 +1,5 @@
 package com.balinasoft.clever.model;
 
 public interface TickHandler {
-    void onTick(long millis);
+    void onTick(int millis);
 }
